requirements.txt
 torch>=2.0.0
Pillow>=9.0.0
numpy>=1.21.0
transformers>=4.30.0
webcolors>=1.12.0 